requirements.txt
nextcord>=2.0.0a6
python-dotenv>=0.19.0